requirements.txt
flask==2.3.2
flask-cors==4.0.0
tensorflow>=2.16.0
werkzeug==2.3.7
numpy==1.24.3
pillow==9.5.0
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0